move `wallet_cli::command::blockchain::{net_sync_faster, find_earliest_epoch, get_last_blockid, util::get_http_peer}` to a new module `exe_common::sync`; delete unused function `wallet_cli::command::blockchain::net_sync_fast`; rustfmt
