+++ exe-common/src/lib.rs
@@ -18,3 +18,4 @@ extern crate tokio_core;
 mod mstream;
 pub mod network;
 pub mod config;
+pub mod sync;
